add vendor profile

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 
@@ -106,11 +107,16 @@ class CategoryController extends Controller
     {
         $category = Category::findOrFail($id);
 
+        $subCategory = SubCategory::where('category_id', $category->id)->count();
+        if($subCategory > 0){
+            return response(['status' => 'error', 'message' =>  'This items contains sub items, for delete this you have to delete the sub items first!']);
+        }
+
+
         $category->delete();
 
-        toastr('Deleted Successfully', 'success');
 
-        return response(['status' => 'success', 'Deleted Successfully!']);
+        return response(['status' => 'success', 'message' =>  'Deleted Successfully!']);
     }
 
     public function changeStatus(Request $request){
